+++ internal/doc/main.go
@@ -53,7 +53,7 @@ import (
 	openshift4_13 "github.com/coreos/butane/config/openshift/v4_13"
 	openshift4_14 "github.com/coreos/butane/config/openshift/v4_14"
 	openshift4_15 "github.com/coreos/butane/config/openshift/v4_15"
-	openshift4_16_exp "github.com/coreos/butane/config/openshift/v4_16_exp"
+	openshift4_16 "github.com/coreos/butane/config/openshift/v4_16"
 	openshift4_8 "github.com/coreos/butane/config/openshift/v4_8"
 	openshift4_9 "github.com/coreos/butane/config/openshift/v4_9"
 	r4e1_0 "github.com/coreos/butane/config/r4e/v1_0"
@@ -124,7 +124,6 @@ func generate(dir string) error {
 			"openshift",
 			[]version{
 				// inverse order of website navbar
-				{"4.16.0-experimental", openshift4_16_exp.Config{}},
 				{"4.8.0", openshift4_8.Config{}},
 				{"4.9.0", openshift4_9.Config{}},
 				{"4.10.0", openshift4_10.Config{}},
@@ -133,6 +132,7 @@ func generate(dir string) error {
 				{"4.13.0", openshift4_13.Config{}},
 				{"4.14.0", openshift4_14.Config{}},
 				{"4.15.0", openshift4_15.Config{}},
+				{"4.16.0", openshift4_16.Config{}},
 			},
 		},
 		{